+++ CHANGELOG.md
@@ -100,6 +100,7 @@ We could not reconstruct _all_ changes, but we tried our best to make the most o
 - [Captcha] Removed ext_scripts/captcha.php, because it was not in use
 - [User] Removed ICQ field in user data
 - [User] Removed MSN field in user data
+- [Tournament] Removed support for WWCL (WWCL shutdown in 2013)
 
 ### Fixed